requirements.txt
fastapi
uvicorn
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.6
faiss-cpu
beautifulsoup4
requests
python-dotenv